[#264] WIP Make assignment history relative to current date (not viewing date) WIP
